+++ tests-smoke/test_admin_email_one_off.py
@@ -1,5 +1,5 @@
+from common import pretty_print
 from dotenv import load_dotenv
-import json
 from notifications_python_client.authentication import create_jwt_token
 import os
 import requests
@@ -16,10 +16,6 @@ ADMIN_CLIENT_SECRET = os.environ.get("ADMIN_CLIENT_SECRET")
 ADMIN_CLIENT_USER_NAME = os.environ.get("ADMIN_CLIENT_USER_NAME")
 EMAIL_TO = os.environ.get("EMAIL_TO")
 API_HOST_NAME = os.environ.get("API_HOST_NAME")
-
-
-def pretty_print(data):
-    print(json.dumps(data, indent=4, sort_keys=True))
 
 
 def test_admin_email_one_off():
